move region dataclass to utils
